Updated load problems tests for names.
